Fix signed -> unsigned bug
I have recently become more aware of how
dangerous using signed and unsigned ints together
is, and I keep introducing bugs while trying to
fix my interfaces to use int instead of some
unsigned value.

+++ src/rfm69_vp.h
@@ -26,7 +26,7 @@ VP_RX_ERROR_T rfm69_vp_rx(
 		rfm69_context_t *rfm,
 		uint8_t *rx_buffer,
 		uint16_t buffer_size,
-		uint16_t timeout_ms,
+		int timeout_ms,
 		int *rssi
 );
